api/requirements.txt
tensorflow>=2.17.0
tensorflow-serving-api>=2.17.0
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
streamlit
